backend/requirements.txt
fastapi
uvicorn
transformers
torch
Pillow
PyPDF2
streamlit
accelerate>=0.26.0
python-multipart
